Create ThrottledTask to throttle status updates in response to file changes
